Added a test case: a factorial function written entirely in csl. This works but I'd prefer a better way to define functions rather than having to manually QUOTE all the expressions that must not be evaluated.
Technically, the DEFUN at this point is closer to a common lisp macro than
a common lisp function. Might not be a bad idea to make this DEFUN a
DEFMACRO and rewrite DEFUN as a special case that automatically quotes
everything in the rest of the lisp before creating the funtion body.